Add a new C launcher for Protégé.
There is seemingly no way to force Protégé to run natively under arm64
if the main executable program of the application is a script, for
whatever reason.

So the first program we execute in the app needs to be a compiled,
binary program with support for both the x86_64 and arm64 architectures.
The only task of the program is to start the actual launch script, which
does all that is needed to start Protégé (including setting the Java
parameters, which are thus still easily modifiable by users).