Start the JVM directly from within the launcher.
Do not start the JVM by exec'ing the java binary. That is the easiest
way to do it, but it does not allow the Java application to receive
Apple events, which are needed to allow users to open OWL files in
Protégé from within the Finder.

Instead, we dynamically load the Java JNI library from the bundled JRE,
and start the JVM in a dedicated thread. The main thread is kept busy in
a dummy loop, whose sole purpose is to allow it to receive events.